// src/client/schema.d.ts
export interface components {
    schemas: {
// ...
        DataPage_SimpleItemSchema_: {
            /** Data */
            data: components["schemas"]["SimpleItemSchema"][];
            /** Total */
            total: number;
            /** Page */
            page: number;
            /** Size */
            size: number;
            /** Pages */
            pages: number;
        };
// ...
        InventorySlotSchema: {
            /**
             * Slot
             * @description Inventory slot identifier.
             */
            slot: number;
            /**
             * Code
             * @description Item code.
             */
            code: string;
            /**
             * Quantity
             * @description Quantity in the slot.
             */
            quantity: number;
        };
        /** ItemResponseSchema */
        ItemResponseSchema: {
            data: components["schemas"]["ItemSchema"];
        };
        /** ItemSchema */
        ItemSchema: {
            /**
             * Name
             * @description Item name.
             */
            name: string;
            /**
             * Code
             * @description Item code. This is the item's unique identifier (ID).
             */
            code: string;
            /**
             * Level
             * @description Item level.
             */
            level: number;
            /**
             * Type
             * @description Item type.
             */
            type: string;
            /**
             * Subtype
             * @description Item subtype.
             */
            subtype: string;
            /**
             * Description
             * @description Item description.
             */
            description: string;
            /**
             * Conditions
             * @description Item conditions. If applicable. Conditions for using or equipping the item.
             */
            conditions?: components["schemas"]["ConditionSchema"][];
            /**
             * Effects
             * @description List of object effects. For equipment, it will include item stats.
             */
            effects?: components["schemas"]["SimpleEffectSchema"][];
            /** @description Craft information. If applicable. */
            craft?: components["schemas"]["CraftSchema"];
            /**
             * Tradeable
             * @description Item tradeable status. A non-tradeable item cannot be exchanged or sold.
             */
            tradeable: boolean;
            /**
             * Recyclable
             * @description Item recyclable status. A recyclable item can be recycled at the matching workshop.
             * @default false
             */
            recyclable: boolean;
        };
// ...
        MonsterSchema: {
            /**
             * Name
             * @description Name of the monster.
             */
            name: string;
            /**
             * Code
             * @description The code of the monster. This is the monster's unique identifier (ID).
             */
            code: string;
            /**
             * Level
             * @description Monster level.
             */
            level: number;
            /** @description Monster type. */
            type: components["schemas"]["MonsterType"];
            /**
             * Hp
             * @description Monster hit points.
             */
            hp: number;
            /**
             * Attack Fire
             * @description Monster fire attack.
             */
            attack_fire: number;
            /**
             * Attack Earth
             * @description Monster earth attack.
             */
            attack_earth: number;
            /**
             * Attack Water
             * @description Monster water attack.
             */
            attack_water: number;
            /**
             * Attack Air
             * @description Monster air attack.
             */
            attack_air: number;
            /**
             * Res Fire
             * @description Monster % fire resistance.
             */
            res_fire: number;
            /**
             * Res Earth
             * @description Monster % earth resistance.
             */
            res_earth: number;
            /**
             * Res Water
             * @description Monster % water resistance.
             */
            res_water: number;
            /**
             * Res Air
             * @description Monster % air resistance.
             */
            res_air: number;
            /**
             * Critical Strike
             * @description Monster % critical strike.
             */
            critical_strike: number;
            /**
             * Initiative
             * @description Monster initiative for turn order.
             */
            initiative: number;
            /**
             * Effects
             * @description List of effects.
             */
            effects?: components["schemas"]["SimpleEffectSchema"][];
            /**
             * Min Gold
             * @description Monster minimum gold drop.
             */
            min_gold: number;
            /**
             * Max Gold
             * @description Monster maximum gold drop.
             */
            max_gold: number;
            /**
             * Drops
             * @description Monster drops. This is a list of items that the monster drops after killing the monster.
             */
            drops: components["schemas"]["DropRateSchema"][];
        };
// ...
        ResourceSchema: {
            /**
             * Name
             * @description The name of the resource
             */
            name: string;
            /**
             * Code
             * @description The code of the resource. This is the resource's unique identifier (ID).
             */
            code: string;
            /**
             * Skill code
             * @description The skill required to gather this resource.
             */
            skill: components["schemas"]["GatheringSkill"];
            /**
             * Level
             * @description The skill level required to gather this resource.
             */
            level: number;
            /**
             * Drops
             * @description The drops of this resource.
             */
            drops: components["schemas"]["DropRateSchema"][];
        };
// ...
        SimpleItemSchema: {
            /**
             * Code
             * @description Item code.
             */
            code: string;
            /**
             * Quantity
             * @description Item quantity.
             */
            quantity: number;
        };
// ...
        StaticDataPage_MonsterSchema_: {
            /** Data */
            data: components["schemas"]["MonsterSchema"][];
            /** Total */
            total: number;
            /** Page */
            page: number;
            /** Size */
            size: number;
            /** Pages */
            pages: number;
        };
// ...
        StaticDataPage_ResourceSchema_: {
            /** Data */
            data: components["schemas"]["ResourceSchema"][];
            /** Total */
            total: number;
            /** Page */
            page: number;
            /** Size */
            size: number;
            /** Pages */
            pages: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}